feat: create DateString component
